+++ server/lib/api/user/index.js
@@ -4,7 +4,7 @@ const bcrypt = require('bcrypt');
 
 const DB = require('../../db');
 const http = require('../../util').http;
-const logger = require('../../logger');
+const exception = require('../../exception');
 
 const saltRounds = 10;
 
@@ -21,17 +21,13 @@ module.exports = {
             'id'
           ));
         } else {
-          res.status(http.codes.NOT_FOUND).send(http.errorFormatter(
-            'That user was not found.'));
+          next(new exception.HttpError('That user was not found.',
+            null, http.codes.NOT_FOUND));
         }
       })
       .catch((err) => {
-
-        const correlationId = logger.log('error', err.stack);
-
-        res.status(http.codes.INTERNAL_SERVER_ERROR)
-          .send(http.errorFormatter(correlationId,
-            'An error occurred looking up this user.'));
+        next(new exception.HttpError('An error occurred finding this user',
+          err, http.codes.INTERNAL_SERVER_ERROR));
       });
   },
 
@@ -54,12 +50,8 @@ module.exports = {
         ));
       })
       .catch((err) => {
-
-        const correlationId = logger.log('error', err.stack);
-
-        res.status(http.codes.INTERNAL_SERVER_ERROR)
-          .send(http.errorFormatter(correlationId,
-            'An error occurred creating this user'));
+        next(new exception.HttpError('An error occurred updating this user',
+          err, http.codes.INTERNAL_SERVER_ERROR));
       });
   },
 
@@ -70,7 +62,7 @@ module.exports = {
       'username'
     );
 
-    DB.user.update(userDetails)
+    DB.user.update(req.params.userId, userDetails)
       .then((user) => {
         res.status(http.codes.OK).send(user.pickLowerCase(
           'email',
@@ -79,17 +71,53 @@ module.exports = {
         ));
       })
       .catch((err) => {
-
-        const correlationId = logger.log('error', err.stack);
-
-        res.status(http.codes.INTERNAL_SERVER_ERROR)
-          .send(http.errorFormatter(correlationId,
-            'An error occurred updating this user'));
+        next(new exception.HttpError('An error occurred updating this user',
+          err, http.codes.INTERNAL_SERVER_ERROR));
       });
   },
 
   updatePassword: (req, res, next) => {
-    res.status(http.codes.NOT_IMPLEMENTED).send();
+
+    const bodyData = req.body.pick('oldPassword', 'newPassword');
+
+    if (bodyData.oldPassword === bodyData.newPassword) {
+      next(new exception.HttpError('Old and new passwords must not match.',
+        null, http.codes.BAD_FORMAT));
+    }
+
+    DB.user.get(req.params.userId)
+      .then((user) => {
+        if (user) {
+          return bcrypt.compare(bodyData.oldPassword, user.password)
+            .then((equal) => {
+
+              if (equal) {
+                return bcrypt.hash(bodyData.newPassword, saltRounds);
+              } else {
+                next(new exception.HttpError('Invalid password.',
+                  null, http.codes.BAD_FORMAT));
+              }
+            })
+            .then((newHash) => {
+              user.password = newHash;
+              return DB.user.update(req.params.userId, user);
+            })
+            .then((updatedUser) => {
+              res.status(http.codes.OK).send(updatedUser.pickLowerCase(
+                'email',
+                'username',
+                'id'
+              ));
+            });
+
+        } else {
+          next(new exception.HttpError('This user was not found.',
+            null, http.codes.NOT_FOUND));
+        }
+      }) .catch((err) => {
+        next(new exception.HttpError('This user was not found.',
+          err, http.codes.NOT_FOUND));
+      });
   },
 
   resetPassword: (req, res, next) => {