+++ server/lib/exception/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  HttpError: require('./http')
+};