+++ server/middleware/index.js
@@ -3,5 +3,6 @@
 module.exports = {
   oauth: require('./oauth'),
   unless: require('./unless'),
-  validation: require('./validation')
+  validation: require('./validation'),
+  error: require('./error')
 };