+++ server/middleware/validation/index.js
@@ -3,24 +3,28 @@
 const v = require('../../lib/validation');
 const http = require('../../lib/util').http;
 const logger = require('../../lib/logger');
+const exception = require('../../lib/exception');
 
-const validator = (req, res, next) => {
+const validator = (path) => {
+  return (req, res, next) => {
 
-  const validationResult = v(req);
+    const validationResult = v(req, path);
 
-  if (validationResult.errors.length) {
+    if (validationResult.errors.length) {
 
-    const correlationId = logger.log('error', 'Received invalid JSON for ' +
-      `path ${req.originalUrl}; JSON body was ${JSON.stringify(req.body)}`);
+      logger.log('error', 'Received invalid JSON for ' +
+        `path ${req.originalUrl}; JSON body was ${JSON.stringify(req.body)}`);
 
-    res.status(http.codes.BAD_FORMAT)
-      .send(http.errorFormatter(correlationId, {
-          validationErrors: validationResult.errors
-            .map((error) => error.message)
-        }));
+      next(new exception.HttpError({
+        validationErrors: validationResult.errors.map((error) => {
+          return error.message;
+        })
+      }, null, http.codes.BAD_FORMAT))
 
-  } else {
-    next();
+    } else {
+      next();
+    }
+
   }
 };