+++ server/routes/index.js
@@ -1,15 +1,15 @@
 'use strict';
 
 const express = require('express');
-const router = express.Router();
+const router = express.Router({mergeParams: true});
 
 /**
  * Mount routes in wrapper function so that we can retain path information.
  *
  * @param router
  */
-require('./user')(router);
-require('./login')(router);
-require('./token')(router);
+
+router.use('/user', require('./user'));
+router.use('/login', require('./login'));
 
 module.exports = router;