+++ server/routes/login/index.js
@@ -1,10 +1,11 @@
 'use strict';
 
+const express = require('express');
+const router = express.Router({mergeParams: true});
+
 const API = require('../../lib/api');
 
-const mount = (router) => {
-  router.route('/login')
-    .post(API.oauth.token);
-};
+router.route('/')
+  .post(API.oauth.token);
 
-module.exports = mount;
+module.exports = router;