+++ server/routes/user/password/index.js
@@ -1,14 +1,21 @@
 'use strict';
 
+const express = require('express');
+const router = express.Router({mergeParams: true});
+
 const API = require('../../../lib/api');
+const middleware = require('../../../middleware');
 
-const mount = (router) => {
-  router.route('/user/:userId/password/reset')
-    .post(API.user.resetPassword);
+router.route('/reset')
+  .post(
+    middleware.validation('/user/password/reset'),
+    API.user.resetPassword
+  );
 
-  router.route('/user/:userId/password/update')
-    .post(API.user.updatePassword);
+router.route('/update')
+  .post(
+    middleware.validation('/user/password/update'),
+    API.user.updatePassword
+  );
 
-};
-
-module.exports = mount;
+module.exports = router;